requirements.txt
django
django-admin-interface
django-crispy-forms
crispy-bootstrap4
pillow
psycopg2-binary
psycopg2
djangorestframework
requests
django-axes
django-admin-confirm
django-simple-captcha
django-multi-captcha-admin
django-recaptcha
django-recaptcha2
cloudinary
whitenoise
xhtml2pdf
django-recaptcha2

